Move ParseSizeAndCSS to modules/html
